+++ assets/scripts/foreground.js
@@ -73,12 +73,4 @@ if ( jQuery( '#ca-addsection' ).length ) {
   // Turn categories into labels
   jQuery('#mw-normal-catlinks ul li a').addClass('label');
 
-  // Make the Page Action button respond to hover
-  //jQuery('button.button').mouseenter(function(){
-  //  jQuery('ul#drop1').addClass('open').addClass('right').css('top', '32px').css('left', '785px');
-  //});
-  //jQuery('ul#drop1').mouseleave(function(){
-  //  jQuery('ul#drop1').removeClass('open').css('top', '-9999px').css('left', '785px');
-  //});
-
 });